Replace include:: with inlined grammar snippets.
The asciidoctor `include::` directive doesn't work everywhere (e.g., in
GitHub renderings), and it also causes extreme brittleness because it
depends on line numbers in the grammar file (which can change a lot!).
Copying snippets verbatim into the documentation is still brittle, but
less so than using external links, and it's clearer both how to spot
differences and how to fix them.