Re-order grammatical elements.
Move comments and whitespace to the beginning of the documentation, as that's actually the first thing to know about creating an (otherwise-empty) file. Put identifiers before expressions, since they have been introduced first within the discussion of values. Leave values before all of the rest of it (i.e., identifiers and expressions), as I want to start from a top-down perspective on the grammar.